Start Idris.Parser.Parser namespace
